Move all the top-level command impls into one file.
